refactor: Common.tsx를 TopInfo에서도 사용하기 위해 한 뎁스 위로 위치 변경
